fixed cmake and changed cast to c++ style and change long long to be the platform independent 64x type

+++ coreengine/settings.cpp
@@ -781,8 +781,8 @@ void Settings::saveSettings()
         settings.endGroup();
 
         settings.beginGroup("Autosaving");
-        settings.setValue("AutoSavingTime",           (long long) autoSavingCylceTime.count());
-        settings.setValue("AutoSavingCycle",                  autoSavingCycle);
+        settings.setValue("AutoSavingTime",            static_cast<qint64>(autoSavingCylceTime.count()));
+        settings.setValue("AutoSavingCycle",           autoSavingCycle);
         settings.endGroup();
 
         // mods
